requirements.txt
PyGithub==1.35